# openapi.yaml
openapi: 3.1.0
info:
  title: Veilence-MX API
  description: >
    Enterprise SaaS supply chain compromise monitor for PyPI and npm packages.
    Provides endpoints for package monitoring, release analysis, alert management,
    organization/RBAC, notifications, and system administration.
  version: 1.0.0
  contact:
    name: Veilence-MX Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.veilence.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens
      description: Exchanges a valid refresh token for a new token pair.
      security: []
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Initiate password reset
      description: >
        Sends a password reset email. Always returns 200 to prevent user enumeration.
      security: []
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: Resets a user's password using a valid reset token.
      security: []
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address
      description: Verifies a user's email using a valid verification token.
      security: []
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── Auth (Protected) ───────────────────────────────────────────
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns the authenticated user's profile.
      operationId: getMe
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Auth]
      summary: Update profile
      description: Updates the authenticated user's first name and last name.
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      description: Changes the authenticated user's password.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/send-verification:
    post:
      tags: [Auth]
      summary: Send verification email
      description: Sends a new email verification token to the authenticated user.
      operationId: sendVerificationEmail
      responses:
        "200":
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/api-keys:
    post:
      tags: [Auth]
      summary: Create API key
      description: Creates a new API key for the authenticated user.
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created (includes the raw key - shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_ApiKeyCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Auth]
      summary: List API keys
      description: Returns all API keys for the authenticated user.
      operationId: listApiKeys
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_ApiKeyList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/api-keys/{id}:
    delete:
      tags: [Auth]
      summary: Revoke API key
      description: Revokes (deletes) an API key belonging to the authenticated user.
      operationId: revokeApiKey
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Sessions ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login or /api/auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key created via /api/auth/api-keys
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: string

    # ── Domain Schemas ──────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        isActive:
          type: boolean
        emailVerified:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiKeyInfo:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        scope:
          type: string
          enum: [read, write, admin]
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
    UpdateProfileRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        scope:
          type: string
          enum: [read, write, admin]
          default: read
        expiresAt:
          type: string
          format: date-time
          description: Optional expiration date (RFC3339)
...
    ApiResponse_Message:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
        error:
          type: string
          nullable: true
...
    ApiResponse_TokenPair:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
        error:
          type: string
          nullable: true
    ApiResponse_User:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"
        error:
          type: string
          nullable: true
...
    ApiResponse_ApiKeyCreated:
      type: object
      properties:
        data:
          type: object
          properties:
            apiKey:
              $ref: "#/components/schemas/ApiKeyInfo"
            key:
              type: string
              description: The raw API key - shown only once at creation time
        error:
          type: string
          nullable: true
    ApiResponse_ApiKeyList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ApiKeyInfo"
        error:
          type: string
          nullable: true
...